% Generated by build script
\name{libio_contributors}
\alias{libio_contributors}
\title{Get contributors for a repository on Libraries.io}
\usage{
libio_contributors()
}
\arguments{
\item{owner}{Repository owner (e.g. "tidyverse")}
\item{repo}{Repository name (e.g. "dplyr")}
\item{api_key}{API key (optional)}
}
\value{tibble: github_id, login, contributions, name, blog}
\description{Get contributors for a repository on Libraries.io}
