% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libio.R
\name{libio_platforms}
\alias{libio_platforms}
\title{List all package manager platforms on Libraries.io}
\usage{
libio_platforms()
}
\value{
tibble: name, project_count, homepage, default_language
}
\description{
List all package manager platforms on Libraries.io
}
